chore: Refactors tests to use Descriptions
Also restructures the extdata files to be numerical in numbering
instead of based on the issue number that it was fixing. The issue number
is now in the description for the test. I added one small test for
replace_na to ensure it works as expected when there are no NAs in the
data frame to replace. inst/extdata/bib2df_testfile_issue_29.bib was
deleted but the contents were moved to bib2df_testfile_1.bib because
they can all be read in together.

Version number was kept the same because there are no functional changes.